+++ interface/rigol.py
@@ -1,0 +1,4 @@
+from interface.gpib import GPIB
+
+class Rigol(GPIB):
+    pass
